+++ extra/constants.js
@@ -1,0 +1,18 @@
+const permissions = {
+    'getUsers1': {
+      all: ['prabal-raghav'],
+      read: ['shivam', 'vinay'],
+      write: ['vinay'],
+      delete: ['prabal-raghav']
+    },
+    'getUsers2': {
+      all: ['prabal-raghav'],
+      read: ['neha', 'vinay'],
+      write: ['vinay'],
+      delete: ['prabal-raghav']
+    }
+  }
+  export {permissions};
+  
+  
+
