--- extra/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "extra",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "pattern:diamond": " node patterns/diamond.js",
-    "pattern:equilateral": "node patterns/equilateral.js"
-     
-  },
-  "author": "",
-  "license": "ISC"
-}
