+++ extra/patterns/index.js
@@ -1,0 +1,3 @@
+import fun  from "./diamond.js";
+import myfun from "./equilateral.js";
+export { fun, myfun};
