+++ extra/utils/validation.js
@@ -1,23 +1,8 @@
+import {validateEmail} from "./helpers.js";
 const validEmail=[];
 const invalidEmail=[];
-const users=[
-  {
-    traineeEmail:"shivam.sharma@successive.tech",
-    reviewerEmail:"megha.rawat@succssive.tech"
-  },
-  {
-    traineeEmail:"neha.goel@succesive.tech",
-    reviewerEmail:"megha.rawat@successive.tech"
 
-  }
-]
-validateEmail=function(email){
-  const regex= /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((successive.tech))$/;
-   return regex.test(email);
-   //console.log("-=-=-=-=-", x)
-   
-}
-validateUsers=function(users){
+const validateUsers=function(users){
 users.forEach((element)=>
    {
        const{traineeEmail,reviewerEmail}=element;
@@ -32,11 +17,13 @@ else
 invalidEmail.push(reviewerEmail);
 
 })
-
-  
 }
-validateUsers(users);
+
+//validateUsers(users);
 console.log( `valid emails: ${validEmail}`);
 console.log("invalid emails: " + invalidEmail);
 console.log("valid count: "+ validEmail.length);
 console.log("invalid count: "+ invalidEmail.length);
+
+export {validateUsers};
+
